fix: use platform.windows cinterop instead of custom header
All Win32 API functions (CreateFileW, GetStdHandle, GetConsoleMode,
SetConsoleMode, GetConsoleScreenBufferInfo, SetConsoleTextAttribute,
GetFileInformationByHandle, GetFileType, GetComputerNameExW) are
already available via platform.windows cinterop. The custom
winapi-util-extras cinterop header caused mingwX64 compilation
failures:

- Struct types (_CONSOLE_SCREEN_BUFFER_INFO, _BY_HANDLE_FILE_INFORMATION)
  were in platform.windows, not our custom cinterop package
- COMPUTER_NAME_FORMAT enum has a private constructor in cinterop,
  making it impossible to construct from Kotlin
- UIntVarOf<UInt>.value needed explicit import of kotlinx.cinterop.value

Removed the custom cinterop header and def file, switched all imports
to platform.windows, and used _COMPUTER_NAME_FORMAT companion
constants (following the pattern from gethostname-kotlin). Used
memScoped+alloc instead of cValue/usePinned for struct field access.

+++ src/mingwMain/kotlin/io/github/kotlinmania/winapiutil/PlatformWin.kt
@@ -4,9 +4,10 @@
 package io.github.kotlinmania.winapiutil
 
 import io.github.kotlinmania.windowssys.windows.win32.foundation.HANDLE
-import platform.windows.INVALID_HANDLE_VALUE
-import io.github.kotlinmania.winapiutil.cinterop.CreateFileW as winCreateFileW
-import io.github.kotlinmania.winapiutil.cinterop.GetStdHandle as winGetStdHandle
+import io.github.kotlinmania.windowssys.windows.win32.foundation.INVALID_HANDLE_VALUE
+import kotlinx.cinterop.toLong
+import platform.windows.CreateFileW as winCreateFileW
+import platform.windows.GetStdHandle as winGetStdHandle
 
 internal const val STD_INPUT_HANDLE: UInt = 0xFFFFFFF6u
 internal const val STD_OUTPUT_HANDLE: UInt = 0xFFFFFFF5u
@@ -29,7 +30,7 @@ internal actual fun openRawHandleForRead(path: String): HANDLE {
             0u,
             null,
         )
-    if (handle?.toLong() == INVALID_HANDLE_VALUE.toLong()) {
+    if (handle?.toLong() == INVALID_HANDLE_VALUE) {
         throw RuntimeException("Failed to open file: $path")
     }
     return handle?.toLong() ?: 0L
@@ -46,7 +47,7 @@ internal actual fun openRawHandleForReadAny(path: String): HANDLE {
             FILE_FLAG_BACKUP_SEMANTICS,
             null,
         )
-    if (handle?.toLong() == INVALID_HANDLE_VALUE.toLong()) {
+    if (handle?.toLong() == INVALID_HANDLE_VALUE) {
         throw RuntimeException("Failed to open file: $path")
     }
     return handle?.toLong() ?: 0L
